% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{isna}
\alias{isna}
\title{Function to count NAs per dataframe column}
\usage{
isna(df)
}
\arguments{
\item{df}{dataframe to check for missing values}
}
\value{
a dataframe with number of missing observations per column
}
\description{
Function to count NAs per dataframe column
}
